Improve catch-all rewriter caching strategy (#36)
#34 introduced a catch-all route handler that acts as a fallback
rewriter for our content pages. It had implementations of caching
mechanisms, but these did not seem to work as expected in production.

This pr introduces the use of `generateStaticParams` to cache route
invocations for the paths present in our `sitemap` on build time, which
results in instant server responses on our rewritten pages.