Improve usage page prose (#62)
This pr improves the sandboxes chart code to be more readable + adds an
indication for the current month (billing period) to be easily
comparable with the other usage charts. Also this pr improves the
general prose of the usage page to be more in sync with what the
components show.