Fix: correctly pass query params into /templates infra openapi-fetch (#77)
This pr fixes correct invocation of `/templates` infra route to get the
correct templates based on the `teamID`. Previously the query param for
`teamID` was passed outside the `params` prop, which made it look like
the query params are passed correctly, but actually they were not
applied to the final invocation url.

This resulted in the templates still being returned, but the team for
which the templates are returned was arbitrarily chosen, based on the
user details inferred from the access token. As a result, users of the
dashboard saw the same templates, regardless of their selected team.

This issue was tricky to spot, since the infra route for `/templates`
accepts this param as nullable and does not throw, but rather return the
templates for an arbitrary team.

Additionally this pr:
- improves the openapi-fetch client definition to explicitly use
next.js's version of fetch instead of globalThis.fetch
- next.js fetch logging by default, which makes issues like these easier
to spot in development
- improves the `VERBOSE` flag condition, for being able to turn of
verbose logging in development environments