Fix: Sandbox list capped at 1000 entries (#117)
The current implementation uses infra api route `/v2/sandboxes` to list
the sandboxes, which uses server side pagination. it did not cause
unexpected behavior yet, since the default limit for entries on a
paginated slice is 1000 sandboxes. Since we officially do not use the
pagination in the dashboard implementation, this basically is a hard cap
at 1000 concurrent sandboxes.

This pr: 
- reverts this change to use the legacy sandbox list endpoint for now
- adds improvements for the initial sandbox metrics fetching