Feature: Add "Last Used" column to API Keys page (#140)
This pr adds a new column, which informs users when an api key was last
used. if the key creation date is before the production collection date
and last_used is undefined, we show "N/A", since we can't know if it was
used or not before the collection date.

### New API Keys Table
<img width="1257" height="408" alt="Screenshot 2025-09-12 at 3 15 42 PM"
src="https://github.com/user-attachments/assets/89378816-5d7e-4905-815c-c1952b6e3b75"
/>

---------

Co-authored-by: Jakub Novák <jakub@e2b.dev>